base: refactor image loading
Moved the image loading into the base classes.
Added new attribute imagefile to store the file name of
the image to be used. In this way the base class takes
care of the image loading within the constructor instead of
taking care of the image loading within each card.
Using os.path.join for building paths.

Added an error message on run_attack_animation if called after
the destruction of the image attribute (by die command) since
it happened to me while testing this patch.
Added some debug code in die method as well for debugging the issue.

Signed-off-by: Sandro Bonazzola <sandro.bonazzola@gmail.com>